Rename @annotation to @index
The API spec and the test suite has been fully updated. In the syntax spec, everything except the prose of the "Data Annotations" section has been updated. I added an issue marker. @msporny, could you please update the prose and the title of that section.

I used "indexes" as the plural of "index" as that seemed to be more commonly used than "indices" according Google :-)

This addresses #216.